Tests(*): move hooks to separate location
