+++ test/lib/types/definition.js
@@ -8,84 +8,127 @@ import {Attribute} from "#@/lib/types/attribute.js";
 import {Filter} from "#@/lib/types/filter.js";
 import {instantiateFromFixture} from "./attribute.js";
 
+// Load data to use in tests from adjacent JSON file
 const basepath = path.relative(process.cwd(), path.dirname(url.fileURLToPath(import.meta.url)));
 const fixtures = fs.readFile(path.join(basepath, "./definition.json"), "utf8").then((f) => JSON.parse(f));
+// Default parameter values to use in tests
 const params = {name: "Test", id: "urn:ietf:params:scim:schemas:Test"};
 
 describe("SCIMMY.Types.SchemaDefinition", () => {
-    it("should require valid 'name' argument at instantiation", () => {
-        assert.throws(() => (new SchemaDefinition()),
-            {name: "TypeError", message: "Required parameter 'name' missing from SchemaDefinition instantiation"},
-            "SchemaDefinition instantiated without 'name' parameter");
-        assert.throws(() => (new SchemaDefinition("")),
-            {name: "TypeError", message: "Expected 'name' to be a non-empty string in SchemaDefinition instantiation"},
-            "SchemaDefinition instantiated with empty string 'name' parameter");
-        assert.throws(() => (new SchemaDefinition(false)),
-            {name: "TypeError", message: "Expected 'name' to be a non-empty string in SchemaDefinition instantiation"},
-            "SchemaDefinition instantiated with 'name' parameter boolean value 'false'");
-        assert.throws(() => (new SchemaDefinition({})),
-            {name: "TypeError", message: "Expected 'name' to be a non-empty string in SchemaDefinition instantiation"},
-            "SchemaDefinition instantiated with complex object 'name' parameter value");
-    });
-    
-    it("should require valid 'id' argument at instantiation", () => {
-        assert.throws(() => (new SchemaDefinition("Test")),
-            {name: "TypeError", message: "Required parameter 'id' missing from SchemaDefinition instantiation"},
-            "SchemaDefinition instantiated without 'id' parameter");
-        assert.throws(() => (new SchemaDefinition("Test", "")),
-            {name: "TypeError", message: "Expected 'id' to be a non-empty string in SchemaDefinition instantiation"},
-            "SchemaDefinition instantiated with empty string 'id' parameter");
-        assert.throws(() => (new SchemaDefinition("Test", false)),
-            {name: "TypeError", message: "Expected 'id' to be a non-empty string in SchemaDefinition instantiation"},
-            "SchemaDefinition instantiated with 'id' parameter boolean value 'false'");
-        assert.throws(() => (new SchemaDefinition("Test", {})),
-            {name: "TypeError", message: "Expected 'id' to be a non-empty string in SchemaDefinition instantiation"},
-            "SchemaDefinition instantiated with complex object 'id' parameter value");
-    });
-    
-    it("should require 'id' to start with 'urn:ietf:params:scim:schemas:' at instantiation", () => {
-        assert.throws(() => (new SchemaDefinition("Test", "test")),
-            {name: "TypeError", message: "Invalid SCIM schema URN namespace 'test' in SchemaDefinition instantiation"},
-            "SchemaDefinition instantiated with invalid 'id' parameter value 'test'");
-    });
-    
-    it("should require valid 'description' argument at instantiation", () => {
-        assert.throws(() => (new SchemaDefinition(...Object.values(params), false)),
-            {name: "TypeError", message: "Expected 'description' to be a string in SchemaDefinition instantiation"},
-            "SchemaDefinition instantiated with 'description' parameter boolean value 'false'");
-        assert.throws(() => (new SchemaDefinition(...Object.values(params), {})),
-            {name: "TypeError", message: "Expected 'description' to be a string in SchemaDefinition instantiation"},
-            "SchemaDefinition instantiated with complex object 'description' parameter value");
-    });
-    
-    it("should have instance member 'name'", () => {
-        assert.strictEqual((new SchemaDefinition(...Object.values(params)))?.name, params.name,
-            "SchemaDefinition did not include instance member 'name'");
-    });
-    
-    it("should have instance member 'id'", () => {
-        assert.strictEqual((new SchemaDefinition(...Object.values(params)))?.id, params.id,
-            "SchemaDefinition did not include instance member 'id'");
-    });
-    
-    it("should have instance member 'description'", () => {
-        assert.ok("description" in (new SchemaDefinition(...Object.values(params))),
-            "SchemaDefinition did not include instance member 'description'");
-    });
-    
-    it("should have instance member 'attributes' that is an array", () => {
-        const definition = new SchemaDefinition(...Object.values(params));
+    describe("@constructor", () => {
+        it("should expect 'name' argument to be defined", () => {
+            assert.throws(() => (new SchemaDefinition()),
+                {name: "TypeError", message: "Required parameter 'name' missing from SchemaDefinition instantiation"},
+                "SchemaDefinition instantiated without 'name' parameter");
+            assert.throws(() => (new SchemaDefinition("")),
+                {name: "TypeError", message: "Expected 'name' to be a non-empty string in SchemaDefinition instantiation"},
+                "SchemaDefinition instantiated with empty string 'name' parameter");
+            assert.throws(() => (new SchemaDefinition(false)),
+                {name: "TypeError", message: "Expected 'name' to be a non-empty string in SchemaDefinition instantiation"},
+                "SchemaDefinition instantiated with 'name' parameter boolean value 'false'");
+            assert.throws(() => (new SchemaDefinition({})),
+                {name: "TypeError", message: "Expected 'name' to be a non-empty string in SchemaDefinition instantiation"},
+                "SchemaDefinition instantiated with complex object 'name' parameter value");
+        });
         
-        assert.ok("attributes" in definition,
-            "SchemaDefinition did not include instance member 'attributes'");
-        assert.ok(Array.isArray(definition.attributes),
-            "SchemaDefinition instance member 'attributes' was not an array");
+        it("should require valid 'id' argument", () => {
+            assert.throws(() => (new SchemaDefinition("Test")),
+                {name: "TypeError", message: "Required parameter 'id' missing from SchemaDefinition instantiation"},
+                "SchemaDefinition instantiated without 'id' parameter");
+            assert.throws(() => (new SchemaDefinition("Test", "")),
+                {name: "TypeError", message: "Expected 'id' to be a non-empty string in SchemaDefinition instantiation"},
+                "SchemaDefinition instantiated with empty string 'id' parameter");
+            assert.throws(() => (new SchemaDefinition("Test", false)),
+                {name: "TypeError", message: "Expected 'id' to be a non-empty string in SchemaDefinition instantiation"},
+                "SchemaDefinition instantiated with 'id' parameter boolean value 'false'");
+            assert.throws(() => (new SchemaDefinition("Test", {})),
+                {name: "TypeError", message: "Expected 'id' to be a non-empty string in SchemaDefinition instantiation"},
+                "SchemaDefinition instantiated with complex object 'id' parameter value");
+        });
+        
+        it("should require 'id' to start with 'urn:ietf:params:scim:schemas:'", () => {
+            assert.throws(() => (new SchemaDefinition("Test", "test")),
+                {name: "TypeError", message: "Invalid SCIM schema URN namespace 'test' in SchemaDefinition instantiation"},
+                "SchemaDefinition instantiated with invalid 'id' parameter value 'test'");
+        });
+        
+        it("should require valid 'description' argument", () => {
+            assert.throws(() => (new SchemaDefinition(...Object.values(params), false)),
+                {name: "TypeError", message: "Expected 'description' to be a string in SchemaDefinition instantiation"},
+                "SchemaDefinition instantiated with 'description' parameter boolean value 'false'");
+            assert.throws(() => (new SchemaDefinition(...Object.values(params), {})),
+                {name: "TypeError", message: "Expected 'description' to be a string in SchemaDefinition instantiation"},
+                "SchemaDefinition instantiated with complex object 'description' parameter value");
+        });
+    });
+    
+    describe("#name", () => {
+        it("should be defined", () => {
+            assert.ok("name" in new SchemaDefinition(...Object.values(params)),
+                "Instance member 'name' was not defined");
+        });
+        
+        it("should be a string", () => {
+            assert.ok(typeof new SchemaDefinition(...Object.values(params))?.name === "string",
+                "Instance member 'name' was not a string");
+        });
+        
+        it("should equal 'name' argument supplied at instantiation", () => {
+            assert.strictEqual(new SchemaDefinition(...Object.values(params))?.name, params.name,
+                "Instance member 'name' did not equal 'name' argument supplied at instantiation");
+        });
+    });
+    
+    describe("#id", () => {
+        it("should be defined", () => {
+            assert.ok("id" in new SchemaDefinition(...Object.values(params)),
+                "Instance member 'id' was not defined");
+        });
+    
+        it("should be a string", () => {
+            assert.ok(typeof new SchemaDefinition(...Object.values(params))?.id === "string",
+                "Instance member 'id' was not a string");
+        });
+        
+        it("should equal 'id' argument supplied at instantiation", () => {
+            assert.strictEqual(new SchemaDefinition(...Object.values(params))?.id, params.id,
+                "Instance member 'id' did not equal 'id' argument supplied at instantiation");
+        });
+    });
+    
+    describe("#description", () => {
+        it("should be defined", () => {
+            assert.ok("description" in new SchemaDefinition(...Object.values(params)),
+                "Instance member 'description' was not defined");
+        });
+    
+        it("should be a string", () => {
+            assert.ok(typeof new SchemaDefinition(...Object.values(params))?.description === "string",
+                "Instance member 'description' was not a string");
+        });
+    
+        it("should equal 'description' argument supplied at instantiation", () => {
+            assert.strictEqual(new SchemaDefinition(...Object.values(params), "Test Description")?.description, "Test Description",
+                "Instance member 'description' did not equal 'description' argument supplied at instantiation");
+        });
+    });
+    
+    describe("#attributes", () => {
+        it("should be defined", () => {
+            assert.ok("attributes" in new SchemaDefinition(...Object.values(params)),
+                "Instance member 'attributes' was not defined");
+        });
+        
+        it("should be an array", () => {
+            assert.ok(Array.isArray(new SchemaDefinition(...Object.values(params))?.attributes),
+                "Instance member 'attributes' was not an array");
+        });
     });
     
     describe("#describe()", () => {
-        it("should have instance method 'describe'", () => {
+        it("should be implemented", () => {
             assert.ok(typeof (new SchemaDefinition(...Object.values(params))).describe === "function",
-                "Instance method 'describe' not defined");
+                "Instance method 'describe' was not defined");
         });
         
         it("should produce valid SCIM schema definition objects", async () => {
@@ -104,9 +147,9 @@ describe("SCIMMY.Types.SchemaDefinition", () => {
     });
     
     describe("#attribute()", () => {
-        it("should have instance method 'attribute'", () => {
+        it("should be implemented", () => {
             assert.ok(typeof (new SchemaDefinition(...Object.values(params))).attribute === "function",
-                "Instance method 'attribute' not defined");
+                "Instance method 'attribute' was not defined");
         });
         
         it("should find attributes by name", () => {
@@ -205,9 +248,9 @@ describe("SCIMMY.Types.SchemaDefinition", () => {
     });
     
     describe("#extend()", () => {
-        it("should have instance method 'extend'", () => {
+        it("should be implemented", () => {
             assert.ok(typeof (new SchemaDefinition(...Object.values(params))).extend === "function",
-                "Instance method 'extend' not defined");
+                "Instance method 'extend' was not defined");
         });
         
         it("should expect 'extension' argument to be an instance of SchemaDefinition or collection of Attribute instances", () => {
@@ -262,9 +305,9 @@ describe("SCIMMY.Types.SchemaDefinition", () => {
     });
     
     describe("#truncate()", () => {
-        it("should have instance method 'truncate'", () => {
+        it("should be implemented", () => {
             assert.ok(typeof (new SchemaDefinition(...Object.values(params))).truncate === "function",
-                "Instance method 'truncate' not defined");
+                "Instance method 'truncate' was not defined");
         });
         
         it("should do nothing without arguments", () => {
@@ -305,6 +348,15 @@ describe("SCIMMY.Types.SchemaDefinition", () => {
                 "Instance method 'truncate' did not remove named attribute directly included in the definition");
         });
         
+        it("should remove named sub-attributes included in the definition", () => {
+            const definition = new SchemaDefinition(...Object.values(params), "", [new Attribute("complex", "test", {}, [new Attribute("string", "test")])]);
+            const expected = JSON.parse(JSON.stringify({...definition.describe(), attributes: [new Attribute("complex", "test").toJSON()]}));
+            const actual = JSON.parse(JSON.stringify(definition.truncate("test.test").describe()));
+            
+            assert.deepStrictEqual(actual, expected,
+                "Instance method 'truncate' did not remove named sub-attribute included in the definition");
+        });
+        
         it("should expect named attributes and sub-attributes to exist", () => {
             const definition = new SchemaDefinition(...Object.values(params));
             
@@ -321,9 +373,9 @@ describe("SCIMMY.Types.SchemaDefinition", () => {
     });
     
     describe("#coerce()", () => {
-        it("should have instance method 'coerce'", () => {
+        it("should be implemented", () => {
             assert.ok(typeof (new SchemaDefinition(...Object.values(params))).coerce === "function",
-                "Instance method 'coerce' not defined");
+                "Instance method 'coerce' was not defined");
         });
         
         it("should expect 'data' argument to be an object", () => {
