+++ test/lib/types/error.js
@@ -2,33 +2,48 @@ import assert from "assert";
 import {SCIMError} from "#@/lib/types/error.js";
 
 describe("SCIMMY.Types.Error", () => {
-    it("should not require arguments at instantiation", () => {
-        assert.doesNotThrow(() => new SCIMError(),
-            "Error type class did not instantiate without arguments");
-    });
-    
     it("should extend native 'Error' class", () => {
         assert.ok(new SCIMError() instanceof Error,
             "Error type class did not extend native 'Error' class");
     });
     
-    it("should have instance member 'name' with value 'SCIMError'", () => {
-        assert.strictEqual((new SCIMError())?.name, "SCIMError",
-            "Error type class did not include instance member 'name' with value 'SCIMError'");
+    describe("@constructor", () => {
+        it("should not require arguments", () => {
+            assert.doesNotThrow(() => new SCIMError(),
+                "Error type class did not instantiate without arguments");
+        });
     });
     
-    it("should have instance member 'status'", () => {
-        assert.ok("status" in (new SCIMError()),
-            "Error type class did not include instance member 'status'");
+    describe("#name", () => {
+        it("should be defined", () => {
+            assert.ok("name" in new SCIMError(),
+                "Instance member 'name' was not defined");
+        });
+        
+        it("should have value 'SCIMError'", () => {
+            assert.strictEqual((new SCIMError())?.name, "SCIMError",
+                "Instance member 'name' did not have value 'SCIMError'");
+        });
     });
     
-    it("should have instance member 'scimType'", () => {
-        assert.ok("scimType" in (new SCIMError()),
-            "Error type class did not include instance member 'scimType'");
+    describe("#status", () => {
+        it("should be defined", () => {
+            assert.ok("status" in new SCIMError(),
+                "Instance member 'status' was not defined");
+        });
     });
     
-    it("should have instance member 'message'", () => {
-        assert.ok("message" in (new SCIMError()),
-            "Error type class did not include instance member 'message'");
+    describe("#scimType", () => {
+        it("should be defined", () => {
+            assert.ok("scimType" in new SCIMError(),
+                "Instance member 'scimType' was not defined");
+        });
+    });
+    
+    describe("#message", () => {
+        it("should be defined", () => {
+            assert.ok("message" in new SCIMError(),
+                "Instance member 'message' was not defined");
+        });
     });
 });
